rough/azure-blob-streamlit-app/requirements.txt
streamlit
azure-storage-blob
python-dotenv